date-parser: handle dates without a year
This is a pretty common case (unfortunately). Try to guess the
appropriate year: take the current year and remove one year if the month
is to great compared to the current month. For example, if we received a
message in March 2015 with a date in February, we assume February
2015. With a date in March or April, we assume April 2015 (because it
may be very late in March and some remote servers may have its clock
off). However, we consider June to be too far away in the future to be
in 2015 and therefore, the date will be June 2014.

Signed-off-by: Vincent Bernat <vincent@bernat.im>